src/pyproject.toml
[tool.poetry]
name = "flask-test"
version = "0.1.0"
description = ""
authors = ["balancy <balancy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
sqlalchemy = "^2.0.32"
pika = "^1.3.2"
celery = "^5.4.0"
pydantic = "^2.8.2"
environs = "^11.0.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
